+++ code/ui/ui_main.c
@@ -23,7 +23,7 @@ This must be the very first function compiled into the .qvm file
 #ifdef _MSC_VER
 __declspec(dllexport)
 #endif
-int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6 ) {
+intptr_t vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6 ) {
 	switch ( command ) {
 	case UI_GETAPIVERSION:
 		return UI_API_VERSION;
